Randm clean up and fix for gcc double concatination error
Fixed so that rival can run from build directory (vpp folder). This
means data had to be copied, this will be fix to include a very down
sized verion of cube and/or use redirect so this is not nessary in the
future. Also added a vcpp project for a splash screen and preloader and
quality assurance exe. Well be fixed later as tesseract broke the
ability for this method to replace the *.bat file. Also fixed minor
build error on gcc libraries.