More organization added in new project system
Using home startin directories to allow users to define a predefine
location to look in before checking the main location. This will be used
like a project system with Start.exe creating the folders structure.

Also Changed vs to reflect a bat file allowing start in directory and
home directory. Also in debug mode it prints a log file to
$home/config/debug/log.txt for debuging purposes.

Also changed bin to allow (when using start.exe) to declare bin/x86
bin/x64 or bin/ so that we can either allow people to use multi platorm
or just define the one they want in bin/

Added in base code for linux (and some mac) to help build the start
project. Will need to add a make file and use sdl so i can have a gui
interface. Will set it up with the same code as rival so it should be
fairly easy to make a modularize shrunken version.